# openapi.json
{
  "components": {
    "schemas": {
      "BatchInferProgress": {
        "description": "Aggregate progress for a batched infer job across all language groups.",
        "properties": {
          "language_groups": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LanguageGroupProgress"
            },
            "description": "Per-language-group progress, keyed by language code.\nUses BTreeMap for deterministic JSON serialization.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "language_groups"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "FileResult": {
        "description": "Result for a single processed file.",
        "properties": {
          "content": {
            "description": "Processed file content (CHAT text or CSV).  Empty string when\n`error` is `Some`.",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "MIME-like content discriminator: `\"chat\"` for CHAT files (default),\n`\"csv\"` for tabular output (e.g. opensmile features)."
          },
          "error": {
            "description": "Human-readable error message if this file failed processing.\n`None` for successfully processed files.",
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "$ref": "#/components/schemas/DisplayPath",
            "description": "Display path for this file: a bare basename (`\"sample.cha\"`) or a\nrelative forward-slash path (`\"PWA/TYO_a1.cha\"`) for directory input.\nBackslashes are normalized to forward slashes on construction."
          },
          "provenance": {
            "description": "Processing provenance extracted from the output CHAT file.\nEach entry records one batchalign3 command that was applied\n(command name, engine version, timestamp). Empty for non-CHAT\noutput or files that failed processing.",
            "items": {
              "$ref": "#/components/schemas/ProvenanceEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "filename"
        ],
        "type": "object"
      },
...
      "JobInfo": {
        "description": "`GET /jobs/{id}` response: job progress.\n\nContains full detail about a single job, including per-file statuses.\nPolled by the CLI to drive progress bars and by the dashboard for\nlive-updating tables.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              }
            ]
          },
          "batch_progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BatchInferProgress",
                "description": "Per-language-group progress for batched text commands (morphotag,\nutseg, translate, coref).  Present only while a batch is in flight."
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/ReleasedCommand",
            "description": "Batchalign command that was submitted (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "description": "ISO 8601 timestamp of when the job reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "completed_files": {
            "description": "Number of files that have finished processing (status `Done`).\nInvariant: `0 <= completed_files <= total_files`.",
            "format": "int64",
            "type": "integer"
          },
          "control_plane": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobControlPlaneInfo",
                "description": "Server control-plane metadata for this job when returned by a server."
              }
            ]
          },
          "current_file": {
            "description": "Filename currently being processed.  `None` when the job is queued\nor has reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "duration_s": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DurationSeconds",
                "description": "Wall-clock duration from first file dispatched to last file finished,\nin seconds.  `None` while the job is still active."
              }
            ]
          },
          "error": {
            "description": "Job-level error message (e.g. worker pool exhaustion, memory gate\ntimeout).  Distinct from per-file errors in `file_statuses`.",
            "type": [
              "string",
              "null"
            ]
          },
          "execution_plan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExecutionPlan",
                "description": "Execution plan describing where and how this job is processed.\nPresent for staged-remote jobs; `None` for local execution."
              }
            ]
          },
          "file_statuses": {
            "description": "Per-file processing status, one entry per submitted file.",
            "items": {
              "$ref": "#/components/schemas/FileStatusEntry"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageSpec",
            "description": "Language specification: a resolved ISO 639-3 code or `\"auto\"`."
          },
          "last_cancelled_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "Wall-clock timestamp of the most recent cancel attempt against\nthis job. `None` until at least one cancel arrives. Denormalized\nfrom the `cancellations` audit table for fast list-view rendering\nwithout a JOIN. See `CancellationRecord` for the full audit\nhistory via `GET /jobs/{id}/cancellations`."
              }
            ]
          },
          "last_cancelled_host": {
            "description": "Caller-reported host of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_reason": {
            "description": "Caller-reported reason of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_source": {
            "description": "Wire-format source string of the most recent cancel attempt\n(`tui`, `api`, `signal`, ...). See `CancelSource`.",
            "type": [
              "string",
              "null"
            ]
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "When the job is deferred before execution, the earliest unix timestamp\nat which it should be retried. `None` when the job is immediately\neligible to run."
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.  Determined\nby the pool at dispatch time.  `None` for queued jobs.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "options": {
            "description": "Typed submitted command options captured at job creation time.\n\nThe dashboard uses this to show the original argument/config payload for\ndebugging and operator review without reconstructing a lossy CLI string."
          },
          "source_dir": {
            "description": "Client's original input directory path, used for display in the\ndashboard and CLI output.  May be empty for content-mode submissions.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "description": "ISO 8601 timestamp of when the job was submitted (e.g.\n`\"2026-01-15T10:00:00Z\"`).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.\nUsed for conflict detection (same `submitted_by` + filename =\nduplicate).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine, resolved from\nTailscale or reverse DNS.  For dashboard display only.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job (immutable after submission).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "options",
          "total_files",
          "completed_files"
        ],
        "type": "object"
      },
...
      "JobResultResponse": {
        "description": "`GET /jobs/{id}/results` response: completed job results.",
        "properties": {
          "files": {
            "description": "Per-file results (empty until the job completes).",
            "items": {
              "$ref": "#/components/schemas/FileResult"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Unique job identifier."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Terminal status of the job."
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "type": "object"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "batchalign3: command-line interface, application, server, worker pool, cache",
    "license": {
      "identifier": "BSD-3-Clause",
      "name": "BSD-3-Clause"
    },
    "title": "batchalign",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/bug-reports": {
      "get": {
        "description": "Bug reports are JSON files written to disk by the validation layer when\nit detects semantic errors (e.g., alignment mismatches, monotonicity\nviolations). This endpoint scans the reports directory by mtime so the\ndashboard can surface recent failures without requiring database queries.",
        "operationId": "list_bug_reports",
        "parameters": [
          {
            "description": "Maximum number of reports to return",
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bug report documents (newest first)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bug report storage could not be read"
          }
        },
        "summary": "Return the most recent bug reports, sorted newest-first.",
        "tags": [
          "bug-reports"
        ]
      }
    },
    "/bug-reports/{id}": {
      "get": {
        "description": "The ID corresponds to the JSON filename on disk without the `.json`\nextension. Returns the raw JSON document as-is so the dashboard can\nrender full diagnostic detail (stack traces, file paths, expected vs.\nactual values).",
        "operationId": "get_bug_report",
        "parameters": [
          {
            "description": "Bug report ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bug report JSON document"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bug report not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bug report storage could not be read"
          }
        },
        "summary": "Return a single bug report by its ID (filename stem).",
        "tags": [
          "bug-reports"
        ]
      }
    },
...
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. The response includes the\nfull staged content of each successful output file. Paths-mode jobs keep\nthe execution-host write, but also preserve a staged copy for download.\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{*filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\nstem-based matching (e.g., requesting `sample.wav` will find `sample.cha`)\nto handle the common case where the output extension differs from the input.",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}